+++ features/__init__.py
@@ -2,4 +2,4 @@ from db.models import DbBaseModel
 
 from features.users.models import *
 from features.recipes.models import *
-from features.image_processing.models import *
+from features.images.models import *
