--- features/image_processing/__init__.py
@@ -1,3 +1,0 @@
-import features.image_processing.router as r
-
-router = r.router
