+++ features/images/__init__.py
@@ -1,0 +1,3 @@
+import features.images.router as r
+
+router = r.router
