+++ features/images/input_models.py
