+++ features/images/operations.py
